requirements.txt
numpy==1.23.5
tensorflow==2.12.0
tensorflow-addons==0.20.0
pillow
tk
scikit-learn
matplotlib